moved clickhouse api to __init__
